Fixed command processor to reject lambda or local function callbacks.
The command processor intentionally keeps weakrefs to all callbacks.
Lambdas and local functions will go out of scope when the caller returns,
unless the caller keeps a reference.  Since the command processor
automatically removes callbacks that go out of scope, this results in
code that just "doesn't work".  To prevent this silent "not working"
case, the command processor now only accepts methods and module
functions and raises ValueError on lambdas and local functions.