Use PySide2 with macOS to avoid 5.15.0 QT QPushButton bug.
Although the system theme does not decorate buttons, it is still
affected by this bug for the play, record stream, record stats,
and zoom buttons.  As long as you click the SVG icons, the button
press registers, but not if you click the background.

This change forces macOS to use PySide2 5.14.2.3 which works
correctly.  We can also deploy the dark theme to macOS!