Fixed race condition when start UI with JLS file argument.
The existing code was always scanning for devices, even when the user
specified a file to load from the command line, or by double-clicking
a JLS file in Windows.  The JLS file load was deferred to the command
queue, but it also issued a device scan.  The UI would load the file,
but the respect the device scan results.

Also load base recording if user selects an annotation file.